Move loading of the deployed contracts into its own file contracts.js
This is to avoid a circular dependency issue between settings.js and utils.js but also sets us up better for improving get setup of contracts in the future
