+++ lib/organization.js
@@ -1,6 +1,7 @@
 "use strict";
 const dopts = require('default-options');
 
+import { getDeployedContracts } from './contracts.js';
 import { getValueFromLogs, requireContract } from './utils.js';
 const Avatar = requireContract('Avatar');
 const Controller = requireContract('Controller');
@@ -11,7 +12,6 @@ const Reputation = requireContract("Reputation");
 const AbsoluteVote = requireContract("AbsoluteVote");
 // import { UpgradeScheme } from   './upgradescheme.js';
 
-import { getDeployedContracts } from './settings.js';
 import { SchemeRegistrar } from './schemeregistrar.js';
 
 const CONTRACT_SCHEMEREGISTRAR = 'SchemeRegistrar';
