+++ lib/settings.js
@@ -1,21 +1,15 @@
 const dopts = require('default-options');
-import { requireContract } from './utils.js';
-
-const GenesisScheme = requireContract("GenesisScheme");
-
-/**
-   * These are uninitialized instances of ExtendTruffleContract,
-   * effectively class factories.
- */
-import { GlobalConstraintRegistrar } from   './globalconstraintregistrar.js';
-import { SchemeRegistrar } from   './schemeregistrar.js';
-import { ContributionReward } from   './contributionreward.js';
-import { AbsoluteVote} from "./absoluteVote.js";
-import { TokenCapGC} from "./tokenCapGC.js";
-import { UpgradeScheme } from   './upgradescheme.js';
 
 // Look for a daostack.json file which can override default settings
-let daostackSettings = require('daostack.json');
+
+/* exported daostackSettings */
+let daostackSettings;
+try {
+  daostackSettings = require('daostack.json');
+  // do stuff
+} catch (ex) {
+  daostackSettings = {};
+}
 
 const defaultSettings = {
   providerUrl: "http://localhost:8545",
@@ -25,74 +19,4 @@ const defaultSettings = {
 
 daostackSettings = dopts(daostackSettings, defaultSettings);
 
-const getDeployedContracts = async function() {
-  /**
-   * These are deployed contract instances represented by their respective Arc
-   * javascript wrappers (ExtendTruffleContract).
-   *
-   * `deployed()` is a static method on each of those classes.
-   **/
-  const contributionReward = await ContributionReward.deployed();
-  const genesisScheme = await GenesisScheme.deployed();
-  const globalConstraintRegistrar = await GlobalConstraintRegistrar.deployed();
-  const schemeRegistrar = await SchemeRegistrar.deployed();
-  const tokenCapGC = await TokenCapGC.deployed();
-  const upgradeScheme = await UpgradeScheme.deployed();
-  const absoluteVote = await AbsoluteVote.deployed();
-
-  /**
-   * `contract` here is an uninitialized instance of ExtendTruffleContract,
-   * basically the class factory.
-   * Calling contract.at() (a static method on the class) will return a
-   * the properly initialized instance of ExtendTruffleContract.
-   */
-  const contracts = {
-    ContributionReward: {
-      contract: ContributionReward,
-      address: contributionReward.address,
-    },
-    GenesisScheme: {
-      contract: GenesisScheme,
-      address: genesisScheme.address,
-    },
-    GlobalConstraintRegistrar: {
-      contract: GlobalConstraintRegistrar,
-      address: globalConstraintRegistrar.address,
-    },
-    SchemeRegistrar: {
-      contract: SchemeRegistrar,
-      address: schemeRegistrar.address,
-    },
-    TokenCapGC: {
-      contract: TokenCapGC,
-      address: tokenCapGC.address,
-    },
-    UpgradeScheme: {
-      contract: UpgradeScheme,
-      address: upgradeScheme.address,
-    },
-    AbsoluteVote: {
-      contract: AbsoluteVote,
-      address: absoluteVote.address,
-    }
-  };
-
-  return {
-    allContracts : contracts,
-    defaultVotingMaching: contract.AbsoluteVote,
-    schemes: [
-      contracts.SchemeRegistrar
-      , contracts.UpgradeScheme
-      , contracts.GlobalConstraintRegistrar
-      , contracts.ContributionReward
-    ],
-    votingMachines: [
-      contracts.AbsoluteVote
-    ],
-    globalConstraints: [
-      contracts.TokenCapGC
-    ]
-  };
-};
-
-export { daostackSettings, getDeployedContracts };
+export { daostackSettings };
