+++ lib/upgradescheme.js
@@ -1,8 +1,8 @@
 "use strict";
 const dopts = require('default-options');
 
+import { getDeployedContracts } from './contracts.js';
 import { getDefaultAccount, ExtendTruffleContract, requireContract, getWeb3 } from './utils.js';
-import { getDeployedContracts } from './settings.js';
 
 const SolidityUpgradeScheme = requireContract("UpgradeScheme");
 const DAOToken = requireContract("DAOToken");
